% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bcov_target.R
\docType{data}
\name{Bcov_target}
\alias{Bcov_target}
\title{Covariate data file of the target dataset when the outcome is binary}
\format{
A dataframe with 1939 rows and 18 columns
\describe{
\item Family ID
\item Individual ID
\item standardized covariate
\item square of the standardized covariate
\item confounder 1
\item confounder 2
\item confounder 3
\item confounder 4
\item confounder 5
\item confounder 6
\item confounder 7
\item confounder 8
\item confounder 9
\item confounder 10
\item confounder 11
\item confounder 12
\item confounder 13
\item confounder 14
}
}
\usage{
Bcov_target
}
\description{
This contains covariate information of the individuals in the target dataset following confounders
}
\keyword{datasets}
