% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bphe_discovery.R
\docType{data}
\name{Bphe_discovery}
\alias{Bphe_discovery}
\title{Phenotype data file of the discovery dataset when the outcome is binary}
\format{
A dataframe with 7916 rows and 3 columns
\describe{
\item Family ID
\item Individual ID
\item phenotype (1=controls, 2=cases)
}
}
\usage{
Bphe_discovery
}
\description{
This contains phenotype information of the individuals in the discovery dataset
}
\keyword{datasets}
