% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bphe_target.R
\docType{data}
\name{Bphe_target}
\alias{Bphe_target}
\title{Phenotype data file of the target dataset when the outcome is binary}
\format{
A dataframe with 1939 rows and 3 columns
\describe{
\item Family ID
\item Individual ID
\item phenotype (0=controls, 1=cases)
}
}
\usage{
Bphe_target
}
\description{
This contains phenotype information of the individuals in the target dataset
}
\keyword{datasets}
