% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyData.R
\docType{data}
\name{DummyData}
\alias{DummyData}
\title{PLINK binary data files
This contains DummyData.fam, DummyData.bim and DummyData.bed files, all in one}
\format{
The data files follow the general PLINK format
}
\usage{
DummyData
}
\description{
PLINK binary data files
This contains DummyData.fam, DummyData.bim and DummyData.bed files, all in one
}
\keyword{datasets}
