% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_binary.R
\name{GWAS_binary}
\alias{GWAS_binary}
\title{GWAS_binary function}
\usage{
GWAS_binary(plink_path, b_file, pheno_file, covar_file, n_confounders, thread)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{pheno_file}{Name (with file extension) of the phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{covar_file}{Name (with file extension) of the covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{n_confounders}{Number of confounding variables in the discovery dataset}

\item{thread}{Number of threads used}
}
\value{
This function will perform GWAS and output
\item{B_trd.sum} GWAS summary statistics file with additive SNP effects
}
\description{
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named B_trd.sum
}
\examples{

}
\keyword{gwas}
