% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_quantitative.R
\name{GWAS_quantitative}
\alias{GWAS_quantitative}
\title{GWAS_quantitative function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named Q_out.trd.sum}
\usage{
GWAS_quantitative(
  plink_path,
  b_file,
  Qphe_discovery,
  Qcov_discovery,
  thread = 20,
  summary_output = "Q_out.trd.sum"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{Qphe_discovery}{Name (with file extension) of the phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{Qcov_discovery}{Name (with file extension) of the covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{thread}{Number of threads used}

\item{summary_output}{Name of the SNP effects of the GWAS summary statistics file specified by the user}
}
\value{
This function will perform GWAS and output
\item{Q_out.trd.sum} GWAS summary statistics file with additive SNP effects
}
\description{
GWAS_quantitative function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named Q_out.trd.sum
}
\examples{
\dontrun{
x <- GWAS_quantitative(plink_path, DummyData, Qphe_discovery, Qcov_discovery)
head(x)
}
}
\keyword{gwas}
