% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWEIS_quantitative.R
\name{GWEIS_quantitative}
\alias{GWEIS_quantitative}
\title{GWEIS_quantitative function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named Q_out.add.sum and interaction SNP effects named Q_out.gxe.sum}
\usage{
GWEIS_quantitative(
  plink_path,
  b_file,
  Qphe_discovery,
  Qcov_discovery,
  thread = 20,
  summary_output = "Q_out"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{Qphe_discovery}{Phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{Qcov_discovery}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{thread}{Number of threads used}

\item{summary_output}{Name (prefix) of the additive or interaction SNP effects of the GWEIS summary statistics file specified by the user}
}
\value{
This function will perform GWEIS and output
\item{Q_out.add.sum} GWEIS summary statistics file with additive SNP effects
\item{Q_out.gxe.sum} GWEIS summary statistics file with interaction SNP effects
}
\description{
GWEIS_quantitative function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named Q_out.add.sum and interaction SNP effects named Q_out.gxe.sum
}
\examples{
\dontrun{ 
x <- GWEIS_quantitative(plink_path, DummyData, Qphe_discovery, Qcov_discovery)
head(x[[1]])
head(x[[2]])
}
}
\keyword{gwies,}
\keyword{gxe}
\keyword{interaction,}
