% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_binary.R
\name{PRS_binary}
\alias{PRS_binary}
\title{PRS_binary function
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_out.trd.sum, B_out.add.sum or B_out.gxe.sum}
\usage{
PRS_binary(
  plink_path,
  b_file,
  summary_input = "B_out.trd.sum",
  summary_output = "B_trd"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{summary_input}{Name of the summary statistics file specified by the user}

\item{summary_output}{Name of the PRS file generated using provided summary statistics file specified by the user}
}
\value{
This function will output
\item{B_trd.sscore} PRSs for each target individual using GWAS additive effects
}
\description{
PRS_binary function
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_out.trd.sum, B_out.add.sum or B_out.gxe.sum
}
\examples{
\dontrun{ 
x <- PRS_binary(plink_path, DummyData)
head(x[[1]])
head(x[[2]])
head(x[[3]])
}
}
\keyword{profile}
\keyword{prs,}
\keyword{scores}
