% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qphe_discovery.R
\docType{data}
\name{Qphe_discovery}
\alias{Qphe_discovery}
\title{Phenotype data file of the discovery dataset when the outcome is quantitative}
\format{
A dataframe with 6426 rows and 3 columns
\describe{
\item Family ID
\item Individual ID
\item phenotype
}
}
\usage{
Qphe_discovery
}
\description{
This contains phenotype information of the individuals in the discovery dataset
}
\keyword{datasets}
