% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qphe_target.R
\docType{data}
\name{Qphe_target}
\alias{Qphe_target}
\title{Phenotype data file of the target dataset when the outcome is quantitative}
\format{
A dataframe with 1579 rows and 3 columns
\describe{
\item Family ID
\item Individual ID
\item phenotype
}
}
\usage{
Qphe_target
}
\description{
This contains phenotype information of the individuals in the target dataset
}
\keyword{datasets}
