% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_regular_binary.R
\name{results_regular_binary}
\alias{results_regular_binary}
\title{results_regular_binary function}
\usage{
results_regular_binary(Bphe_target, Bcov_target, n_confounders)
}
\arguments{
\item{n_confounders}{Number of confounding variables in the target dataset}
}
\value{
This function will output
\item{Individual_risk_values_without_confounders.txt} Risk values for each target individual using when there are no confounders
\item{Individual_risk_values.txt} Risk values for each target individual using when there are confounders
}
\description{
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\keyword{disease}
\keyword{individual}
\keyword{risk}
\keyword{risk,}
\keyword{scores,}
\keyword{values,}
