% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_permuted_binary.R
\name{summary_permuted_binary}
\alias{summary_permuted_binary}
\title{summary_permuted_binary function}
\usage{
summary_permuted_binary(Bphe_target, Bcov_target, n_confounders)
}
\arguments{
\item{n_confounders}{Number of confounding variables in the target dataset}
}
\value{
This function will output
\item{Bsummary.txt} the summary of the fitted model
}
\description{
This function uses plink2 and outputs the p value of permuted model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\keyword{regression}
\keyword{summary}
