% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_regular_quantitative.R
\name{summary_regular_quantitative}
\alias{summary_regular_quantitative}
\title{summary_regular_quantitative function
This function uses plink2 and outputs the summary of regular model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_trd.sum, Q_add.sum and Q_gxe.sum}
\usage{
summary_regular_quantitative(
  Qphe_target,
  Qcov_target,
  trd_score = "Q_trd.sscore",
  add_score = "Q_add.sscore",
  gxe_score = "Q_gxe.sscore",
  Model,
  summary_output = "Qsummary.txt",
  risk_output = "Individual_risk_values.txt"
)
}
\arguments{
\item{Qphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Qcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{trd_score}{The .sscore file generated using additive SNP effects of GWAS summary statistics}

\item{add_score}{The .sscore file generated using additive SNP effects of GWEIS summary statistics}

\item{gxe_score}{The .sscore file generated using interaction SNP effects of GWEIS summary statistics}

\item{Model}{Specify the model number (1: y = PRS_trd + E + PRS_trd x E + confounders, 2: y = PRS_add + E + PRS_add x E + confounders, 3: y = PRS_add + E + PRS_gxe x E + confounders, 4: y = PRS_add + E + PRS_gxe + PRS_gxe x E + confounders, where y is the outcome variable, E is the covariate of interest, PRS_trd and PRS_add are the polygenic risk scores computed using additive SNP effects of GWAS and GWEIS summary statistics respectively, and PRS_gxe is the polygenic risk scores computed using GxE interaction SNP effects of GWEIS summary statistics.)}

\item{summary_output}{Name of the model summary file specified by the user}

\item{risk_output}{Name of the file containing risk scores of the target individuals specified by the user}
}
\value{
This function will output
\item{Qsummary.txt} the summary of the fitted model
\item{Individual_risk_values.txt} the estimated risk values of individuals in the target sample
}
\description{
summary_regular_quantitative function
This function uses plink2 and outputs the summary of regular model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_trd.sum, Q_add.sum and Q_gxe.sum
}
\examples{
\dontrun{
x <- summary_regular_quantitative(Qphe_target, Qcov_target, Model = 4)
x[[1]][[1]]
x[[1]][[2]]
x[[1]][[3]]
x[[1]][[4]]
x[[1]][[5]]
x[[1]][[6]]
x[[1]][[7]]
x[[1]][[8]]
x[[1]][[9]]
x[[1]][[10]]
x[[1]][[11]]
x[[1]][[12]]
head(x[[2]])
}
}
\keyword{regression}
\keyword{risk}
\keyword{scores}
\keyword{summary}
