% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_binary.R
\name{GWAS_binary}
\alias{GWAS_binary}
\title{GWAS_binary function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named B_out.trd.sum}
\usage{
GWAS_binary(
  plink_path,
  b_file,
  Bphe_discovery,
  Bcov_discovery,
  thread = 20,
  summary_output = "B_out.trd.sum"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{Bphe_discovery}{Name (with file extension) of the phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{Bcov_discovery}{Name (with file extension) of the covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{thread}{Number of threads used}

\item{summary_output}{Name of the SNP effects of the GWAS summary statistics file specified by the user}
}
\value{
This function will perform GWAS and output
\item{B_out.trd.sum} GWAS summary statistics file with additive SNP effects
}
\description{
GWAS_binary function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named B_out.trd.sum
}
\keyword{gwas}
