% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWEIS_binary.R
\name{GWEIS_binary}
\alias{GWEIS_binary}
\title{GWEIS_binary function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named B_out.add.sum and interaction SNP effects named B_out.gxe.sum}
\usage{
GWEIS_binary(
  plink_path,
  b_file,
  Bphe_discovery,
  Bcov_discovery,
  thread = 20,
  summary_output = "B_out"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{Bphe_discovery}{Phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{Bcov_discovery}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{thread}{Number of threads used}

\item{summary_output}{Name (prefix) of the additive or interaction SNP effects of the GWEIS summary statistics file specified by the user}
}
\value{
This function will perform GWEIS and output
\item{B_out.add.sum} GWEIS summary statistics file with additive SNP effects
\item{B_out.gxe.sum} GWEIS summary statistics file with interaction SNP effects
}
\description{
GWEIS_binary function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named B_out.add.sum and interaction SNP effects named B_out.gxe.sum
}
\keyword{gwies,}
\keyword{gxe}
\keyword{interaction,}
