% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWEIS_binary.R
\name{GWEIS_binary}
\alias{GWEIS_binary}
\title{GWEIS_binary function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named B_add.sum and interaction SNP effects named B_gxe.sum}
\usage{
GWEIS_binary(
  plink_path,
  b_file,
  pheno_file,
  covar_file,
  n_confounders,
  thread,
  summary_output1 = NULL,
  summary_output2 = NULL
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{pheno_file}{Phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{covar_file}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{n_confounders}{Number of confounding variables in the discovery dataset}

\item{thread}{Number of threads used}
}
\value{
This function will perform GWEIS and output
\item{B_add.sum} GWEIS summary statistics file with additive SNP effects
\item{B_gxe.sum} GWEIS summary statistics file with interaction SNP effects
}
\description{
GWEIS_binary function
This function performs GWEIS using plink2 and outputs the GWEIS summary statistics files with additive SNP effects named B_add.sum and interaction SNP effects named B_gxe.sum
}
\keyword{gwies,}
\keyword{gxe}
\keyword{interaction,}
