% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_binary.R
\name{PRS_binary}
\alias{PRS_binary}
\title{PRS_binary function}
\usage{
PRS_binary(plink_path, b_file)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}
}
\value{
This function will output
\item{B_trd.sscore} PRSs for each target individual using GWAS additive effects
\item{B_add.sscore} PRSs for each target individual using GWEIS additive effects
\item{B_gxe.sscore} PRSs for each target individual using GWEIS interaction effects
}
\description{
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\keyword{profile}
\keyword{prs,}
\keyword{scores}
