% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_regular_binary.R
\name{summary_regular_binary}
\alias{summary_regular_binary}
\title{summary_regular_binary function}
\usage{
summary_regular_binary(Bphe_target, Bcov_target, n_confounders)
}
\arguments{
\item{Bphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Bcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{n_confounders}{Number of confounding variables in the target dataset}
}
\value{
This function will output
\item{Bsummary.txt} the summary of the fitted model
}
\description{
This function uses plink2 and outputs the summary of regular model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\keyword{regression}
\keyword{summary}
